// packages-adapters/bootstrap/src/interfaces.ts
//import { IDragNDropConfig } from '@erp/toolkit';

export type Colors =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'danger'
  | 'warning'
  | 'info'
  | 'light'
  | 'dark'
  | 'body'
  | 'muted'
  | 'white'
  | 'black-50'
  | 'white-50';

export type Variants = 'primary' | 'secondary' | 'success' | 'danger' | 'warning' | 'info' | 'light' | 'dark';

export type BgColors =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'danger'
  | 'warning'
  | 'info'
  | 'light'
  | 'dark'
  | 'white'
  | 'transparent';

export type Sizes = 'sm' | 'lg';

export type Animations = 'glow' | 'wave';

/*export interface IListGroupItem {
    id: number;
    name: string;
    dragconfig?: IDragNDropConfig;
    disabled?: boolean;
    badge?: number;
    icon?: string;
}*/